Added text parsing
Created a class called 'text_parser.py' to allow given text to be parsed into an array which contains only key words for the elastic search.

Input is given to the parse method, which will recursively iterate through the input to normalize and check for spelling, then compare each word against a list of stop words and remove any that are not needed.

Modified the requirements.txt to require the autocorrect framework which works with the parser to potentially fix the spelling of words

From: [OOMCKBOT-20]